docker-compose: Restucture files
This adds a dev version of the file, which mounts the code as a volume,
so changes to the code take effect without a rebuild. To prevent having
to specify -f multiple times on the commandline, create two top-level
main compose files that include others, and hide those others in the
new docker-compose directory.