legacy-convert: Copy location from packet to observation
This just copies the GPS location from each data packet into the
resulting observation objects. This is the same as how the current
backend handles this, but not perfect: Only once every 24h, the GPS
location is refreshed. Also, if the GPS is not working, a zero position
is transmitted for the next 24h, but the previous position could still
be considered valid.

But this ensures that most observations will at least have a location
that is usually correct.