Lower start_interval for frost-db
Previously raising interval meant that on startup, the service would not
be healthy until the second health check after a minute, which is way
too long. By setting start_interval, during startup the health is
checked more often. This also requires setting start_period, which is
apparently the time that any failures will be ignored (not counted
against retries).